Move LLVM-dependent functions in CPPUtil.h to LLVMUtil.h
