+++ lib/CFL/CFLAlias.cpp
@@ -28,7 +28,7 @@
  */
 
 #include "CFL/CFLAlias.h"
-#include "SVF-LLVM/CPPUtil.h"//WHSTODO: Remove me
+#include "Util/CPPUtil.h"//WHSTODO: Remove me
 
 using namespace SVF;
 using namespace cppUtil;
