+++ lib/MemoryModel/PointerAnalysis.cpp
@@ -27,7 +27,7 @@
  *      Author: Yulei Sui
  */
 
-#include "SVF-LLVM/CPPUtil.h"//WHSTODO: remove
+#include "Util/CPPUtil.h"//WHSTODO: remove
 #include "Util/Options.h"
 #include "SVFIR/SVFModule.h"
 #include "Util/SVFUtil.h"
